Update the integration of the consensus/network layers
Addresses #243, #244, and #245.

Instead of using `Chain block` to store the current chain, we use the storage
layer, namely, `Ouroboros.Storage.ChainDB` (as it is not ready yet, a
preliminary version of its API is currently used).

The `ChainDB`:

* Stores blocks: blocks part of the current chain, but also downloaded blocks
  that are not yet part of the current chain or that will only ever be part of
  forks.
* Does chain selection: the responsibility of using `OuroborosTag` to select
  the current chain is shifted to the `ChainDB`. Its API might have to be
  updated to reflect this, i.e, add a `p`(rotocol) type parameter.
* Maintains the ledger state.

We use `ChainFragment`s (containing max k blocks) instead of `Chain`s to store
the current chain and chain candidates. The `OuroborosTag` type class was
updated to reflect this. `ChainFragment`s are compared by the `BlockNo` of the
their head (= the length of the whole chain), instead of the length of the
`ChainFragment` (always <= k).

Header/body split: instead of always using/storing/downloading whole blocks,
headers are now validated before the whole blocks. The last k headers can be
kept in-memory, but not the last k blocks. The old code often assumed `hdr ~
blk`, which is no longer true. The new Block Fetch logic is used (via
`BlockFetchConsensusInterface`) to download blocks in the background.

The ChainSync client was split off in a separate module and updated.

TODO the ChainSync server still has to be rewritten

TODO the tests have to be updated. We will need a (mock) implementation of the
`ChainDB` for this.